Remove duplicated SvgFeatureRenderer code from @jbrowse
