+++ jbrowse/src/client/JBrowse/Browser/plugins/ExtendedVariantPlugin/ExtendedVariantRenderer/index.ts
@@ -1,2 +1,2 @@
-export { default as ReactComponent } from './components/SvgFeatureRendering'
+export { default as ReactComponent } from './components/ExtendedVariantRenderer'
 export { default as configSchema } from './configSchema'
